Begin step 1 edit the hero
Hydrate the heroes component with data

In this situation, our heroes component now has state and is a container ui
with a container html. If we ever have anthing that needs presentational
logic, or business logic we would create more components for that.
But the ask for tour of heroes is such that there is no need for splitting
I also believe that the hero interface belongs

change shared folder to the name shareable

Use uppercase pipe on title